+++ CHANGELOG.md
@@ -22,6 +22,7 @@ This document outlines major changes between releases.
 
 ### Removed
 - `owner_private_key` field from the `issue-secret` output, it duplicated `secret_access_key` (#1344)
+- Unusable "none" option for session-tokens authmate parameter (#1355)
 
 ### Upgrading from 0.46.0